Updated app.js
Completed the foundations of the Game Class, the CardSet Object, and to instantiate he Game class.  When Start Game is pressed, a new game is instantiated, with a new deck of cards, new board, and the random solution based on a random room, weapon, and character is posted to the updating status display (for now).